Larger restructuring concerning native-code binding.
Previously, in EPMC we had a command-line option to decide whether
graph-based computations should be performed in Java or in native code.
In the latter case, graph structures to be used for such computations
were not stored in memory using Java arrays, but rather using
ByteBuffers. Now, the command-line option has been removed and graphs
are always constructed using Java arrays. Because of this, it might
happen that the JVM decides to copy arrays when calling to native code.
As far as I checked, this seems however only to happen for smaller
arrays. The disadvantage of potential copying is probably worth it,
because by avoiding having to decide between arrays and ByteBuffers, we
remove one annoying source of errors.

To still be able to decide between using Java and native implementation,
the source file containing generally used value iteration routines has
been split into a Java and a native part, between which can be chosen
using a command-line option.

To make the Java-based value iterations actually work, I also fixed a
bug when set-ting one array to itself.